라우터 사용에 따른 디렉토리 정리
