+++ src/component/LinkItems/LinkItems.css
@@ -25,13 +25,13 @@
   margin: 8px 20px;
 }
 
-a img {
+.LinkItems a img {
   width: 100%;
   height: 200px;
   transition: transform 0.3s ease;
 }
 
-a img:hover {
+.LinkItems a img:hover {
   transform: scale(1.3);
 }
