+++ src/pages/HomePage/HomePage.js
@@ -1,0 +1,13 @@
+import { Link } from "react-router-dom";
+
+const HomePage = () => {
+  return (
+    <>
+      <Link to="/shared">/shared</Link>
+      <br />
+      <Link to="/folder">/folder</Link>
+    </>
+  );
+};
+
+export default HomePage;
